+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Plugins: Remove edit-calendar, tinymce, json-column, pretty-json-column, translation, email-table, dump-php, master-slave
 - Plugins: Do not fail with a fatal error when a plugin is not an object (bug TimWolla/docker-adminer#211)
 - Plugin login-password-less: Allow the plugin password to be also the password of some server
+- New plugin: Display images in select, same as in Adminer Editor
 
 ## Adminer 5.5.1 (released 2026-07-21)
 - Ignore invalid X-Forwarded-Prefix (GHSA-fr74-9mf9-gf44)